[GH-28202]: Added GetGroupsByNames API (#33558)
* feat: Added GetGroupsByNames API
This commit implements the endpoint discussed in issue #28202.

This adds a new API endpoint to get multiple groups by a list of
names.

Previously, when the app received a post with @ mentions that it
didn't recognize, it would attempt to fetch them all as users,
then if some were still missing, it would go one by one attempting
to fetch each as a group. Now we just fetch all the groups at
once, just like we do for users.

Also added unit tests for the new API and it's respective
documentation.

* Added server version to GetGroupsByNames documentation

Co-authored-by: Harrison Healey <harrisonmhealey@gmail.com>

* fix: updated status_profile_polling tests to use new endpoint

* fix: fixed mock test

Was using get for post request

---------

Co-authored-by: Harrison Healey <harrisonmhealey@gmail.com>